+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This release contains enhancements and bug fixes.
 
 - [#138](https://github.com/blaqkube/mysql-operator/pull/138) Print `Phase` to as part of `kubectl get` columns
 - [#143](https://github.com/blaqkube/mysql-operator/pull/143) Improve the documentation
+- [#144](https://github.com/blaqkube/mysql-operator/issue/144) Add a maintenance window
 
 #### Bug Fixes: